akshare/requirements.txt
akshare==1.8.11
PyYAML==6.0
PyQt5==5.15.7